# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BhuSetu",
    "description": "BhuSetu: Intelligent Land Record Digitization and Validation System for SIH Problem Statement 26018 (Dept of Land Resources, Govt of India)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/records/export": {
      "get": {
        "tags": [
          "Records & Export"
        ],
        "summary": "Export Land Records",
        "description": "Exports digitized land records in JSON, CSV, or GeoJSON formats.\nComplies with DoLR data sharing protocols: automatically masks PII for unauthorized users.",
        "operationId": "export_land_records_api_records_export_get",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv|geojson)$",
              "description": "Export format: json, csv, or geojson",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format: json, csv, or geojson"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (e.g. published, needs_review)",
              "title": "Status"
            },
            "description": "Filter by status (e.g. published, needs_review)"
          },
          {
            "name": "doc_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by doc_type",
              "title": "Doc Type"
            },
            "description": "Filter by doc_type"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/records/export": {
      "get": {
        "tags": [
          "Records & Export"
        ],
        "summary": "Export Land Records",
        "description": "Exports digitized land records in JSON, CSV, or GeoJSON formats.\nComplies with DoLR data sharing protocols: automatically masks PII for unauthorized users.",
        "operationId": "export_land_records_api_v1_records_export_get",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(json|csv|geojson)$",
              "description": "Export format: json, csv, or geojson",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format: json, csv, or geojson"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (e.g. published, needs_review)",
              "title": "Status"
            },
            "description": "Filter by status (e.g. published, needs_review)"
          },
          {
            "name": "doc_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by doc_type",
              "title": "Doc Type"
            },
            "description": "Filter by doc_type"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}